refactor source layout and reverse dockerfile/script dependencies
The initial commit is completely re-factored:
1.  The root build files--Dockerfile and Makefile--are at top level
2.  All supporting scripts are in src/tools
3.  All resources injected into image are in src/resources
4.  ci directory is reserved for later integration of CI/CD
5.  All Caddy source identification is in CADDY_COORDINATES, not just
version.
6.  ReadMe file added, closing issue; but needs more detail

Most importantly, the Docker build drives the work rather than having
a script that prepares the source prior to running the Docker build
process.  This means that all the build can happen within a Docker
context, or outside the Docker build (see the ReadMe file).  The
Dockerfile only contains logic for the image; the details of preparing
the source and compiling are in scripts that are independent of the
Docker build.  This will also facilitate the concourse.ci process using
the docker-image resource (that defaults to using a top-level
Dockerfile).